formatted lists to work w dropdowns
