+++ src/pages/student/RegisterStudent.js
@@ -4,7 +4,7 @@ import ReactGA from 'react-ga';
 import Dropdown from '../../components/layout/dropdown/Dropdown.js'
 import './RegisterStudent.scss';
 import { setFormDetails } from '../../redux/actions/catalog';
-
+import {minorsList, majorsList, customStyles, years} from './DropdownLists'
 import cutelilbearohyesyouare from '../assets/signup.png';
 import illeniumisourlordandsavior from '../assets/login.png';
 import { Fab } from '@material-ui/core';
@@ -19,319 +19,34 @@ const RegisterStudent = ({
   const [showMinorsDD, setShowMinorsDD] = useState(false);
   const [showTagsDD, setShowTagsDD] = useState(false);
   const [isLoggedIn, setIsLoggedIn] = useState(false);
-  
-  const customStyles = {
-    multiValue: (provided, state) => ({
-      ...provided,
-      background: '#D1D3D4',
-      color: '#2b2b2b',
-      borderRadius: 4,
-    }),
-    control: (provided, state) => ({
-      display: 'flex',
-      width: 320,
-      margin: 7,
-      marginBottom: 8,
-      fontSize: 12,
-      fontFamily: 'Roboto, sans-serif',
-      fontWeight: 400,
-      fontStyle: 'normal',
-      borderRadius: 5,
-      border: 'solid 1px #949494',
-      // border: (state.selectProps.error) ? 'solid 1px #ff2d2d' : 'solid 1px #949494',
-    }),
-    menu: (provided, state) => ({
-      ...provided,
-      margin: 8,
-      marginTop: 2,
-      width: 320,
-      fontSize: '12px',
-      fontFamily: 'Qanelas Soft',
-      fontWeight: 300,
-      fontStyle: 'normal',
-      textAlign: 'left',
-      color:
-        state.selectProps.value && state.selectProps.value.length >= 3
-          ? '#cccccc'
-          : '#4e4e4e',
-    }),
-    multiValueRemove: (provided, state) => ({
-      ...provided,
-      background: '#D1D3D4',
-      color: '#2b2b2b',
-      borderRadius: 10,
-      '&:hover': {
-        color: 'hsl(0,0%,40%)',
-      },
-    }),
-    singleValue: (provided, state) => ({
-      ...provided,
-      color: '#4e4e4e',
-    }),
-    multiValueLabel: (provided, state) => ({
-      ...provided,
-      marginLeft: '4px',
-      padding: '2px',
-      paddingLeft: '5px',
-      fontSize: '12px',
-    }),
-    indicatorSeparator: (provided, state) => ({
-      ...provided,
-      width: 0,
-    }),
-
-    clearIndicator: (provided, state) => ({
-      ...provided,
-      cursor: 'pointer',
-    }),
-    dropdownIndicator: (provided, state) => ({
-      ...provided,
-      cursor: 'pointer',
-    }),
-    valueContainer: (provided, state) => ({
-      ...provided,
-      padding: '5px 8px',
-    }),
-    '@media only screen and (min-width: 1700px)': {
-      menu: (provided, state) => ({
-        ...provided,
-        width: 500,
-      }),
-    },
-  };
-  const years = ['Freshman', 'Sophomore', 'Junior', 'Senior'];
-  var yearSelected = false;
   const [year, setYear] = useState('');
-
-  const majorsList = [
-    'African American Studies',
-    'American Studies',
-    'Ancient Egyptian and Near Eastern Art and Archaeology',
-    'Anthropology',
-    'Applied Mathematics',
-    'Architecture',
-    'Art History',
-    'Art Practice',
-    'Asian American and Asian Diaspora Studies',
-    'Astrophysics',
-    'Atmospheric Science',
-    'Bioengineering',
-    'Business Administration',
-    'Celtic Studies',
-    'Chemical Biology',
-    'Chemical Engineering',
-    'Chemistry',
-    'Chicana/o and Latina/o Studies',
-    'Chinese Language',
-    'Civil Engineering',
-    'Civil Engineering',
-    'Classical Civilizations',
-    'Classical Languages',
-    'Cognitive Science',
-    'Comparative Literature',
-    'Computer Science',
-    'Conservation and Resource Studies',
-    'Dance and Performance Studies',
-    'Data Science',
-    'Dutch Studies',
-    'EECS',
-    'East Asian Religion, Thought and Culture',
-    'East European/Eurasian Languages and/or Cultures',
-    'Economics',
-    'Ecosystem Management and Forestry',
-    'Electrical Engineering & Computer Sciences (EECS)',
-    'Energy Engineering',
-    'Engineering Mathematics and Statistics',
-    'Engineering Physics',
-    'English',
-    'Environmental Earth Science',
-    'Environmental Economics and Policy',
-    'Environmental Engineering Science',
-    'Environmental Sciences',
-    'Ethnic Studies',
-    'Film',
-    'French',
-    "Gender and Women's Studies",
-    'Genetics and Plant Biology',
-    'Geography',
-    'Geology',
-    'Geophysics',
-    'German',
-    'Global Studies',
-    'Greek',
-    'History',
-    'Industrial Engineering & Operations Research',
-    'Integrative Biology',
-    'Interdisciplinary Studies',
-    'Italian Studies',
-    'Japanese Language',
-    'Landscape Architecture',
-    'Latin',
-    'Legal Studies',
-    'Linguistics',
-    'Marine Science',
-    'Materials Science & Engineering',
-    'Mathematics',
-    'Mechanical Engineering',
-    'Media Studies',
-    'Microbial Biology',
-    'Molecular Environmental Biology',
-    'Molecular and Cell Biology',
-    'Music',
-    'Native American Studies',
-    'Near Eastern Civilizations',
-    'Near Eastern Languages & Literatures',
-    'Nuclear Engineering',
-    'Nutritional Sciences: Dietetics',
-    'Nutritional Sciences: Physiology & Metabolism',
-    'Nutritional Sciences: Toxicology',
-    'Operations Research & Management Science',
-    'Philosophy',
-    'Physics',
-    'Planetary Science',
-    'Political Economy',
-    'Political Science',
-    'Pre-Med / Pre-Health',
-    'Prelaw ',
-    'Psychology',
-    'Public Health',
-    'Rhetoric',
-    'Scandinavian',
-    'Slavic Languages and Literatures',
-    'Social Welfare',
-    'Society and Environment',
-    'Sociology',
-    'South and Southeast Asian Studies',
-    'Spanish and Portuguese',
-    'Statistics',
-    'Sustainable Environmental Design',
-    'Theater and Performance Studies',
-    'Urban Studies',
-  ];
   const [majorsPopulated, setMajorsPopulated] = useState(0);
+
+  var yearSelected = false;
+
+  const formattedYears = years.map((year) => {
+    return {value: year, label: year}
+  })
+
+  const formattedMajors = majorsList.map((major) => {
+    return {value: major, label: major}
+  });
+
+  const formattedMinors = minorsList.map((minor) => {
+    return {value: minor, label: minor}
+  });
+
   var selectedMajors = [];
-
-  const minorsList = [
-    'African American Studies',
-    'American Studies',
-    'Ancient Egyptian and Near Eastern Civilizations',
-    'Anthropology',
-    'Applied Language Studies',
-    'Arabic',
-    'Architecture',
-    'Armenian Studies',
-    'Art History',
-    'Asian American and Asian Diaspora Studies',
-    'Astrophysics',
-    'Atmospheric Science',
-    'Bioengineering',
-    'Buddhist Studies',
-    'Celtic Studies',
-    'Chemical Engineering',
-    'Chemistry',
-    'Chicana/o and Latina/o Studies',
-    'Chinese Language',
-    'City Planning',
-    'Classical Civilizations',
-    'Climate Science',
-    'Comparative Literature',
-    'Computer Science',
-    'Conservation and Resource Studies',
-    'Creative Writing',
-    'Dance and Performance Studies',
-    'Data Science',
-    'Demography',
-    'Digital Humanities',
-    'Disability Studies',
-    'Dutch Studies',
-    'Earth and Planetary Science',
-    'East European/Eurasian Languages and/or Cultures',
-    'Ecosystem Management and Forestry',
-    'Education',
-    'Electrical Engineering & Computer Sciences (EECS)',
-    'Electronic Intelligent Systems',
-    'Energy Engineering',
-    'Energy and Resources',
-    'English',
-    'Environmental Design & Urbanism in Developing Countries',
-    'Environmental Earth Science',
-    'Environmental Economics and Policy',
-    'Environmental Engineering',
-    'Ethnic Studies',
-    'Food Systems',
-    'French',
-    "Gender and Women's Studies",
-    'Geography',
-    'Geology',
-    'Geophysics',
-    'Geospatial Information Science and Technology',
-    'Geosystems',
-    'German',
-    'Global Poverty and Practice',
-    'Global Studies',
-    'Greek',
-    'Hebrew',
-    'Hispanic Languages, Linguistics, and Bilingualism',
-    'History',
-    'History of the Built Environment',
-    'Human Rights Interdisciplinary',
-    'Industrial Engineering & Operations Research',
-    'Italian Studies',
-    'Jewish Studies',
-    'Journalism',
-    'Korean Language',
-    'Landscape Architecture',
-    'Latin',
-    'Lesbian, Gay, Bisexual, & Transgender Studies',
-    'Linguistics',
-    'Logic',
-    'Marine Science',
-    'Materials Science & Engineering',
-    'Mathematics',
-    'Mechanical Engineering',
-    'Medieval Studies',
-    'Music',
-    'Native American Studies',
-    'Nuclear Engineering',
-    'Nutritional Sciences: Physiology & Metabolism',
-    'Nutritional Sciences: Toxicology',
-    'Persian',
-    'Philosophy',
-    'Physics',
-    'Planetary Science',
-    'Political Economy',
-    'Portuguese Language, Literatures, & Cultures',
-    'Psychology',
-    'Public Health',
-    'Public Policy',
-    'Race and the Law',
-    'Rhetoric',
-    'Russian Culture',
-    'Russian Language',
-    'Russian Literature',
-    'Scandinavian',
-    'Science and Math Education',
-    'Social & Cultural Factors in Environmental Design',
-    'South and Southeast Asian Studies',
-    'Spanish Languages, Literatures, & Cultures',
-    'Statistics',
-    'Structural Engineering',
-    'Sustainable Design',
-    'Theater and Performance Studies',
-    'Tibetan',
-    'Turkish',
-  ];
   var selectedMinors = [];
-
   var selectedTags = [];
-
   const { tags } = formDetails;
+
   function setSelectedYear(value) {
     // Set profile values.
     setYear(value);
     yearSelected = value != null;
-    console.log('Selecting ' + value);
-    console.log(yearSelected);
+    //console.log('Selecting ' + value);
+    //console.log(yearSelected);
   }
 
   function addSelectedMajor(major) {
@@ -345,7 +60,7 @@ const RegisterStudent = ({
       selectedMajors.push(major);
     }
 
-    console.log('Majors: ' + selectedMajors);
+    //console.log('Majors: ' + selectedMajors);
   }
 
   function addSelectedMinor(minor) {
@@ -359,7 +74,7 @@ const RegisterStudent = ({
       selectedMinors.push(minor);
     }
 
-    console.log('Minors: ' + selectedMinors);
+    //console.log('Minors: ' + selectedMinors);
   }
 
   function addSelectedTag(tag) {
@@ -373,10 +88,10 @@ const RegisterStudent = ({
       selectedTags.push(tag);
     }
 
-    console.log('Tags: ' + selectedTags);
+    //console.log('Tags: ' + selectedTags);
   }
-  console.log(tagOptions)
-  console.log(years)
+  //console.log(tagOptions)
+  //console.log(years)
 
   ReactGA.initialize('UA-176775736-1');
   ReactGA.pageview('/register');
@@ -412,7 +127,7 @@ const RegisterStudent = ({
 
             {/* Year Dropdown */}
             <Dropdown
-              options = {years}
+              options = {formattedYears}
               multi = {false}
               search = {false}
               placeholder ='Select year'
@@ -420,39 +135,10 @@ const RegisterStudent = ({
               set = {setSelectedYear}
               style={customStyles}
             />
-            {/* <div onClick={() => setShowYearDD(!showYearDD)} class="filter">
-              {yearSelected ? (
-                <div class="selected"> {year} </div>
-              ) : (
-                'Select year'
-              )} */}
-              {/*{!selectedYear ? "Select year" : selectedYear}*/}
-              {/* {showYearDD ? (
-                <i className="fas fa-caret-up"></i>
-              ) : (
-                <i className="fas fa-caret-down"></i>
-              )} }
-              </div>*/}
-           {/*  {showYearDD && (
-              <div className="filter-dropdown-years scrollable-content">
-                {years.map((year) => (
-                  <div className="filter-selection">
-                    <input
-                      type="radio"
-                      name="year"
-                      checked={years[year]}
-                      onClick={() => setSelectedYear(year)}
-                    />
-                    <div class="spacer" />
-                    {year}
-                  </div>
-                ))}
-              </div>
-            )} */}
 
             {/* Majors Dropdown */}
             <Dropdown
-              options = {majorsList}
+              options = {formattedMajors}
               multi = {true}
               search = {true}
               placeholder ='Select major(s)'
@@ -460,37 +146,10 @@ const RegisterStudent = ({
               set = {addSelectedMajor}
               style={customStyles}
             />
-            {/* <div onClick={() => setShowMajorDD(!showMajorsDD)} class="filter">
-              {majorsPopulated ? (
-                <div class="selected"> {selectedMajors.toString()} </div>
-              ) : (
-                'Select major(s)'
-              )}
-              {showMajorsDD ? (
-                <i className="fas fa-caret-up"></i>
-              ) : (
-                <i className="fas fa-caret-down"></i>
-              )}
-            </div>
-            {showMajorsDD && (
-              <div className="filter-dropdown-majors scrollable-content">
-                {majorsList.map((major) => (
-                  <div className="filter-selection">
-                    <input
-                      type="checkbox"
-                      checked={majorsList[major]}
-                      onClick={() => addSelectedMajor(major)}
-                    />
-                    <div class="spacer" />
-                    {major}
-                  </div>
-                ))}
-              </div>
-            )} */}
 
             {/* Minors Dropdown */}
             <Dropdown
-              options = {minorsList}
+              options = {formattedMinors}
               multi = {true}
               search = {true}
               placeholder ='Select minor(s)'
@@ -498,30 +157,7 @@ const RegisterStudent = ({
               set = {addSelectedMinor}
               style={customStyles}
             />
-            {/* <div onClick={() => setShowMinorsDD(!showMinorsDD)} class="filter">
-              Select minor(s)
-              {showMinorsDD ? (
-                <i className="fas fa-caret-up"></i>
-              ) : (
-                <i className="fas fa-caret-down"></i>
-              )}
-            </div>
-            {showMinorsDD && (
-              <div className="filter-dropdown-minors scrollable-content">
-                {minorsList.map((minor) => (
-                  <div className="filter-selection">
-                    <input
-                      type="checkbox"
-                      checked={minorsList[minor]}
-                      onClick={() => addSelectedMinor(minor)}
-                    />
-                    <div class="spacer" />
-                    {minor}
-                  </div>
-                ))}
-              </div>
-            )} */}
-
+            
             {/* Tags Dropdown */}
             <Dropdown
               options = {tagOptions}
@@ -532,34 +168,6 @@ const RegisterStudent = ({
               set = {addSelectedTag}
               style={customStyles}
             />
-
-            {/* <div>
-              <div
-                onClick={() => setShowTagsDD(!showTagsDD)}
-                className={`${showTagsDD && 'openDD'} filter tags-filter`}>
-                Tags
-                {showTagsDD ? (
-                  <i className="fas fa-caret-up"></i>
-                ) : (
-                  <i className="fas fa-caret-down"></i>
-                )}
-              </div>
-              {showTagsDD && (
-                <div className="filter-dropdown-tags scrollable-content">
-                  {tagOptions.map((tag) => (
-                    <div className="filter-selection" key={tag.value}>
-                      <input
-                        type="checkbox"
-                        checked={tags[tag.label]}
-                        onClick={() => addSelectedTag(tag)}
-                      />
-                      <div class="spacer" />
-                      {tag.label}
-                    </div>
-                  ))}
-                </div>
-              )}
-            </div> */}
 
             <div onClick class="save">
               Save
